move every route to /(routes) folder
