+++ src/app/(routes)/@admin/(.)admin/page.tsx
