+++ src/app/(routes)/account/page.tsx
@@ -1,8 +1,8 @@
 "use client";
-import Navbar from "../_components/Navbar";
+import Navbar from "../../_components/Navbar";
 import Image from "next/image";
-import Sidebar from "../_components/account/Sidebar";
-import { useUserContext } from "../_lib/_context/UserContext";
+import Sidebar from "../../_components/account/Sidebar";
+import { useUserContext } from "../../_lib/_context/UserContext";
 import { useRef, useState } from "react";
 export default function AccountPage() {
   const user = useUserContext();
