+++ src/app/(routes)/layout.tsx
@@ -1,7 +1,7 @@
 import type { Metadata } from "next";
 import { Inter } from "next/font/google";
-import { UserProvider } from "./_lib/_context/UserContext";
-import "./globals.css";
+import { UserProvider } from "../_lib/_context/UserContext";
+import "../globals.css";
 
 const inter = Inter({ subsets: ["latin"] });
