+++ src/app/(routes)/login/page.tsx
@@ -1,4 +1,4 @@
-import LoginForm from "../_components/auth/LoginForm";
+import LoginForm from "@/app/_components/auth/LoginForm";
 
 export default function Login() {
   return (
