+++ src/app/(routes)/terms-of-service/page.tsx
