+++ src/app/_components/Navbar.tsx
@@ -1,7 +1,5 @@
 "use client";
-
 import Link from "next/link";
-
 import Image from "next/image";
 import { useState } from "react";
