+++ src/app/_components/auth/SignupForm.tsx
@@ -1,6 +1,6 @@
 import RegisterForm from "../RegisterForm";
-import { registerFormData } from "@/app/_data/RegisterFormData";
-import { registerFormInputs } from "@/app/_data/RegisterFormData";
+import { registerFormInputs } from "@/app/_lib/_data/RegisterFormData";
+import { registerFormData } from "@/app/_lib/_data/RegisterFormData";
 import registerServerAction from "@/app/_actions/registerAction";
 import RegisterFooter from "./RegisterFooter";
